requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
crontab==0.23.0
dparse==0.5.1
idna==3.3
packaging==21.3
pkginfo==1.8.2
pyparsing==3.0.9
PyYAML==6.0
requests==2.27.1
safety==1.10.3
toml==0.10.2
urllib3==1.26.9
